move map data into data folder
